requirements.txt
pyarrow
pandas
apache-airflow[postgres,slack,celery]
dotenv
python-dotenv
fastparquet